Doctor list
Remove your custom jQuery code from previous exercises.

Add jQuery code that will change the background color of a h1 element
when clicked.

Make all paragraphs have a font size of 18px when they are double
clicked.

Set all li text color to red when the mouse is hovering, reset to white
when it is not.